Refactor StreamService and remove unused NTIMarket methods
StreamService was split into a base interface and a new MutableStreamService for mutable operations. Unused methods from NTIMarketRepository and service implementations were removed to simplify code and improve maintainability.
