+++ apps/backend/src/main/java/net/akarmanov/projectplace/repos/NtiMarketRepository.java
@@ -6,5 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 public interface NtiMarketRepository extends JpaRepository<NTIMarket, UUID> {
-  NTIMarket findByName(String name);
 }
