+++ apps/backend/src/main/java/net/akarmanov/projectplace/services/nti/NtiMarketService.java
@@ -6,11 +6,8 @@ import java.util.List;
 import java.util.UUID;
 
 public interface NtiMarketService {
-  void checkNtiMarket(UUID ntiMarketId);
 
   NTIMarket getNtiMarket(UUID ntiMarketId);
-
-  List<NTIMarket> getNtiMarkets();
 
   List<NTIMarket> getNtiMarkets(List<UUID> ids);
 }
