+++ apps/backend/src/main/java/net/akarmanov/projectplace/services/stream/DefaultStreamService.java
@@ -6,14 +6,11 @@ import net.akarmanov.projectplace.domain.NTIMarket;
 import net.akarmanov.projectplace.domain.Stream;
 import net.akarmanov.projectplace.repos.NtiMarketRepository;
 import net.akarmanov.projectplace.repos.StreamRepository;
-import net.akarmanov.projectplace.services.exceptions.StreamImageUploadException;
 import net.akarmanov.projectplace.services.exceptions.StreamNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 import java.util.UUID;
@@ -25,17 +22,11 @@ import static org.springframework.data.jpa.domain.Specification.where;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class StreamServiceImpl implements StreamService {
+public class DefaultStreamService implements StreamService {
 
-  private final StreamRepository streamRepository;
+  protected final StreamRepository streamRepository;
 
-  private final NtiMarketRepository ntiMarketRepository;
-
-  @Override
-  @PreAuthorize("hasRole('ADMIN')")
-  public Stream create(Stream createdStream) {
-    return streamRepository.save(createdStream);
-  }
+  protected final NtiMarketRepository ntiMarketRepository;
 
   @Override
   public Stream getById(UUID id) {
@@ -49,17 +40,6 @@ public class StreamServiceImpl implements StreamService {
   }
 
   @Override
-  public Stream save(Stream stream) {
-    return streamRepository.save(stream);
-  }
-
-  @Override
-  @PreAuthorize("hasRole('SUPER_ADMIN')")
-  public void delete(UUID streamId) {
-    streamRepository.deleteById(streamId);
-  }
-
-  @Override
   public Page<Stream> findAll(Specification<Stream> specification, Pageable pageable) {
     return streamRepository.findAll(specification, pageable);
   }
@@ -67,17 +47,6 @@ public class StreamServiceImpl implements StreamService {
   @Override
   public List<NTIMarket> getNTIMarkets() {
     return ntiMarketRepository.findAll();
-  }
-
-  @Override
-  public void addImage(UUID streamId, MultipartFile file) {
-    var stream = getById(streamId);
-    try {
-      stream.setImageBytes(file.getBytes());
-      streamRepository.save(stream);
-    } catch (Exception e) {
-      throw new StreamImageUploadException(e);
-    }
   }
 
   @Override
