+++ apps/backend/src/main/java/net/akarmanov/projectplace/usecases/StreamUseCase.java
@@ -7,7 +7,7 @@ import net.akarmanov.projectplace.mapping.NtiMarketMapper;
 import net.akarmanov.projectplace.mapping.StreamMapper;
 import net.akarmanov.projectplace.rest.api.dto.NTIMarketDto;
 import net.akarmanov.projectplace.rest.api.dto.StreamDto;
-import net.akarmanov.projectplace.services.stream.StreamService;
+import net.akarmanov.projectplace.services.stream.MutableStreamService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
@@ -20,7 +20,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class StreamUseCase {
 
-  private final StreamService streamService;
+  private final MutableStreamService streamService;
 
   private final StreamMapper streamMapper;
